rename precision_plugin.py -> precision.py
