+++ src/lightning/pytorch/plugins/__init__.py
@@ -9,7 +9,7 @@ from lightning.pytorch.plugins.precision.deepspeed import DeepSpeedPrecisionPlug
 from lightning.pytorch.plugins.precision.double import DoublePrecisionPlugin
 from lightning.pytorch.plugins.precision.fsdp import FSDPMixedPrecisionPlugin, FSDPPrecisionPlugin
 from lightning.pytorch.plugins.precision.half import HalfPrecisionPlugin
-from lightning.pytorch.plugins.precision.precision_plugin import PrecisionPlugin
+from lightning.pytorch.plugins.precision.precision import PrecisionPlugin
 from lightning.pytorch.plugins.precision.transformer_engine import TransformerEnginePrecisionPlugin
 from lightning.pytorch.plugins.precision.xla import XLAPrecisionPlugin
