+++ src/lightning/pytorch/plugins/precision/half.py
@@ -20,7 +20,7 @@ from torch import Tensor
 from torch.nn import Module
 
 from lightning.fabric.plugins.precision.utils import _convert_fp_tensor, _DtypeContextManager
-from lightning.pytorch.plugins.precision.precision_plugin import PrecisionPlugin
+from lightning.pytorch.plugins.precision.precision import PrecisionPlugin
 
 
 class HalfPrecisionPlugin(PrecisionPlugin):
