+++ tests/tests_pytorch/accelerators/test_cpu.py
@@ -10,7 +10,7 @@ from lightning.fabric.plugins import TorchCheckpointIO
 from lightning.pytorch import Trainer
 from lightning.pytorch.accelerators import CPUAccelerator
 from lightning.pytorch.demos.boring_classes import BoringModel
-from lightning.pytorch.plugins.precision.precision_plugin import PrecisionPlugin
+from lightning.pytorch.plugins.precision.precision import PrecisionPlugin
 from lightning.pytorch.strategies import SingleDeviceStrategy
 
 from tests_pytorch.helpers.runif import RunIf
